requirements.txt
openai
langchain
langchain_community
python-dotenv
pandas
tiktoken
chromadb
langchain-chroma
langchain-openai